+++ src/main/java/sai_adapa/projs/inv_management/model/users/Users.java
@@ -32,7 +32,7 @@ public class Users implements Serializable {
     private String passwdHash;
     private String sessionToken;
     @Column(unique = true)
-    private Long ratingId;
+    private Integer ratingId;
 
 }
